Fixes some bad ideas from the last commit and some performance boost.
I've removed the textual representation of XMLValidator instances I added
in the last commit. It seemed like a good idea to have it but the fact
is that there is no concrete use-case to justify the extra lines of
code.

The results of XMLValidator.validate and XMLValidator.validate_style are
now being cached.